refactor: 추천제품 페이지 리팩토링
